docker_bs/test_project/requirements.txt
Django==4.1.2
django-notifications-hq==1.7.0
django-redis==5.2.0
django-reversion==5.0.3
django-simpleui==2022.7.29
django-tinymce==3.4.0
ipython==7.13.0
kombu==5.2.4
mysqlclient==2.1.0
Pillow==7.0.0
PyJWT==1.7.1
pymongo==4.2.0
PyMySQL==1.0.2
redis==3.5.3
requests==2.22.0
simplejson==3.16.0
sqlparse==0.4.2
urllib3==1.25.8
uWSGI==2.0.20
xadmin-py3==2.0.4
